# docs/openapi.yaml
openapi: 3.1.0
info:
  title: CTF Full V1 API
  version: 0.1.0
  description: >
    Partial contract skeleton for VS01-VS05 and the read-only External Reality
    Interface. All paths below are relative to the /api/v1 server URL. The runnable
    FastAPI application's generated /openapi.json is authoritative where this
    skeleton does not yet declare an implemented endpoint or exact response shape.
    Domain implementations must expand schemas without weakening ownership, version,
    evidence, genealogy, idempotency, or human-authority requirements. Every
    consequential mutation requires Idempotency-Key in the deployed policy and
    may return 413 or 429 with a stable error envelope.
...
servers:
  - url: /api/v1
security:
  - bearerAuth: []
...
paths:
...
  /projects/{projectId}/ai/execute:
    parameters: [{$ref: "#/components/parameters/ProjectId"}]
    post:
      tags: [AI]
      operationId: executeRegisteredAiOperation
      parameters: [{$ref: "#/components/parameters/IdempotencyKey"}]
      description: >
        Executes only a registered prompt operation. Input is persisted before
        provider invocation. Output remains proposed/candidate and optional
        persistence uses normal domain validation.
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/AIExecuteRequest"}
      responses:
        "200":
          description: Structured candidate output and complete run record
          content:
            application/json:
              schema: {$ref: "#/components/schemas/AIExecutionResult"}
        "403": {$ref: "#/components/responses/ErrorResponse"}
        "422": {$ref: "#/components/responses/ErrorResponse"}
        "429": {$ref: "#/components/responses/RateLimited"}
        "503": {$ref: "#/components/responses/ErrorResponse"}
  /projects/{projectId}/ai/runs:
    parameters: [{$ref: "#/components/parameters/ProjectId"}]
    get:
      tags: [AI]
      operationId: listProjectAiRuns
      responses:
        "200":
          description: Tenant-authorized safe AI execution records
          content:
            application/json:
              schema: {type: array, items: {$ref: "#/components/schemas/AIRun"}}
...
  /eri/khal/assets:
    get:
      tags: [ERI]
      operationId: listKhalAssets
      description: Read-only, tenant-mapped normalized asset query with verified-cache degradation.
      responses:
        "200":
          description: Normalized assets or honest unconfigured state
          content:
            application/json:
              schema: {$ref: "#/components/schemas/KhalReadResponse"}
        "503": {$ref: "#/components/responses/ErrorResponse"}
...
  /eri/khal/measurements:
    get:
      tags: [ERI]
      operationId: queryKhalMeasurements
      description: Read-only, tenant-mapped normalized measurements. Measurement never implies causal attribution.
      parameters:
        - {name: external_metric, in: query, required: true, schema: {type: string}}
        - {name: observed_from, in: query, required: false, schema: {type: string, format: date-time}}
        - {name: observed_to, in: query, required: false, schema: {type: string, format: date-time}}
      responses:
        "200":
          description: Normalized measurements with provenance, quality, and freshness state
          content:
            application/json:
              schema: {$ref: "#/components/schemas/KhalReadResponse"}
        "503": {$ref: "#/components/responses/ErrorResponse"}
  /projects/{projectId}/eri/khal/measurements/ingest:
    parameters: [{$ref: "#/components/parameters/ProjectId"}]
    post:
      tags: [ERI]
      operationId: ingestKhalMeasurements
      description: Pull normalized read-only KHAL measurements into deduplicated RealityEvents without causal attribution.
      parameters:
        - {name: external_metric, in: query, required: true, schema: {type: string}}
        - {name: observed_from, in: query, required: false, schema: {type: string, format: date-time}}
        - {name: observed_to, in: query, required: false, schema: {type: string, format: date-time}}
      responses:
        "201": {description: Normalized RealityEvents and duplicate count}
        "503": {$ref: "#/components/responses/ErrorResponse"}
  /eri/reality-events:
    post:
      tags: [ERI]
      operationId: ingestRealityEvent
      security:
        - providerCredential: []
      parameters: [{$ref: "#/components/parameters/IdempotencyKey"}]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/RealityEvent"}
      responses:
        "201": {description: Validated candidate; not automatically a CTF fact}
        "200": {description: Existing deduplicated event}
components:
  securitySchemes:
    bearerAuth: {type: apiKey, in: header, name: X-Session-Token}
    providerCredential: {type: apiKey, in: header, name: X-Provider-Credential}
  parameters:
    ProjectId: {name: projectId, in: path, required: true, schema: {type: string, format: uuid}}
    IdempotencyKey: {name: Idempotency-Key, in: header, required: true, schema: {type: string, minLength: 1, maxLength: 200}}
...
  responses:
...
    RateLimited:
      description: Per-tenant/session rate limit or tenant AI quota exhausted
      headers:
        Retry-After: {schema: {type: integer, minimum: 1}, description: Seconds before retry}
      content: {application/json: {schema: {$ref: "#/components/schemas/Error"}}}
    ErrorResponse:
      description: Stable safe error
      content: {application/json: {schema: {$ref: "#/components/schemas/Error"}}}
  schemas:
...
    AIExecuteRequest:
      type: object
      required: [operation, user_input]
      properties:
        operation: {type: string}
        user_input: {type: string, minLength: 1, maxLength: 50000}
        consequentiality: {enum: [LOW, MEDIUM, HIGH, CRITICAL], default: MEDIUM}
        prompt_version: {type: [string, "null"]}
        context: {type: object}
        persist_as: {type: [string, "null"]}
        expected_version: {type: [integer, "null"]}
...
    AIRun:
      type: object
      required: [id, project_id, operation, provider, model, prompt_id, prompt_version, methodology_version, outcome, retry_count]
      properties:
        id: {type: string}
        project_id: {type: string}
        operation: {type: string}
        provider: {type: string}
        model: {type: string}
        prompt_id: {type: string}
        prompt_version: {type: string}
        methodology_version: {type: string}
        outcome: {enum: [STARTED, SUCCEEDED, FAILED]}
        input_tokens: {type: integer, minimum: 0}
        cached_input_tokens: {type: integer, minimum: 0}
        output_tokens: {type: integer, minimum: 0}
        latency_ms: {type: integer, minimum: 0}
        estimated_cost_usd: {type: string}
        price_snapshot_id: {type: string}
        retry_count: {type: integer, minimum: 0, maximum: 1}
        error: {type: [object, "null"]}
    AIExecutionResult:
      type: object
      required: [run, output]
      properties:
        run: {$ref: "#/components/schemas/AIRun"}
        output:
          type: object
          required: [status, items]
          properties:
            status: {enum: [PROPOSED, CANDIDATE]}
            items: {type: array, items: {type: object}}
            summary: {type: string}
...
    RealityEvent:
      type: object
      required: [provider, external_event_id, event_type, observed_at, subject]
      properties:
        provider: {type: string}
        external_event_id: {type: string}
        event_type: {type: string}
        observed_at: {type: string, format: date-time}
        generated_at: {type: string, format: date-time}
        subject: {type: object}
        metric: {type: string}
        baseline: {type: object}
        observed: {type: object}
        data_quality: {enum: [VALID, ESTIMATED, MISSING, STALE, OUTLIER, INVALID, UNKNOWN]}
        received_at: {type: string, format: date-time}
        source_provenance: {type: object}
        transformation_history: {type: array, items: {type: object}}
...
    KhalReadResponse:
      type: object
      required: [items, read_only, status]
      properties:
        items: {type: array, items: {type: object}}
        read_only: {const: true}
        status: {enum: [ADAPTER_NOT_CONFIGURED, READY, DEGRADED]}
        last_success_at: {type: [string, "null"], format: date-time}
        stale: {type: boolean}
        age_seconds: {type: number, minimum: 0}
        source: {enum: [LAST_VERIFIED_CACHE]}
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, request_id]
          properties:
            code: {type: string}
            message: {type: string}
            request_id: {type: string}